src/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
chromadb==0.4.15
librosa==0.10.1
soundfile==0.12.1
audioread==3.0.1
scikit-learn==1.4.0
numpy==1.26.4
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
pandas==2.2.0
matplotlib==3.8.2
scipy==1.12.0
joblib>=1.3.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Dimensionality Reduction
umap-learn==0.5.5

# Dashboard
streamlit==1.31.0
plotly==5.18.0